Fix README typos and security issues: add server-side auth, fix TTS HTTPS

+++ radio.ino
@@ -39,6 +39,10 @@ unsigned long lastMQTTPublish = 0;
 
 #define MAX_STATIONS 20
 
+// Service menu authentication
+const char* SERVICE_USER = "admin";
+const char* SERVICE_PASS = "jolka";
+
 Audio audio;
 WebServer server(80);
 Preferences preferences;
@@ -100,7 +104,7 @@ String urlEncode(String str) {
 volatile bool isSpeaking = false;
 
 void speak(String text, int volume = 21) {
-  String ttsURL = "http://translate.google.com/translate_tts?ie=UTF-8&client=tw-ob&tl=pl&q=" + 
+  String ttsURL = "https://translate.google.com/translate_tts?ie=UTF-8&client=tw-ob&tl=pl&q=" + 
                   urlEncode(text);
   
   Serial.println("Speaking: " + text);
@@ -145,7 +149,7 @@ void speakIP(String ip) {
   
   String messageEN = "Connected. I P address: " + ipTextEN;
   
-  String ttsURL = "http://translate.google.com/translate_tts?ie=UTF-8&client=tw-ob&tl=en&q=" + 
+  String ttsURL = "https://translate.google.com/translate_tts?ie=UTF-8&client=tw-ob&tl=en&q=" + 
                   urlEncode(messageEN);
   
   Serial.println("Speaking (EN): Connected. IP address: " + ip);
@@ -751,12 +755,7 @@ const char index_html[] PROGMEM = R"rawliteral(
     };
 
     var openServiceMenu = function() {
-      var password = prompt(currentLang === 'pl' ? 'Podaj hasło serwisowe:' : 'Enter service password:');
-      if (password === 'jolka') {
-        window.location.href = '/service';
-      } else if (password !== null) {
-        alert(currentLang === 'pl' ? 'Nieprawidłowe hasło!' : 'Incorrect password!');
-      }
+      window.location.href = '/service';
     };
 
     var stationsData = [];
@@ -1293,10 +1292,16 @@ void handleRoot() {
 }
 
 void handleService() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   server.send_P(200, "text/html", service_html);
 }
 
 void handleGetPins() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   String json = "{";
   json += "\"sda\":" + String(SDA_PIN) + ",";
   json += "\"scl\":" + String(SCL_PIN) + ",";
@@ -1310,6 +1315,9 @@ void handleGetPins() {
 }
 
 void handleGetMQTT() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   String json = "{";
   json += "\"server\":\"" + MQTT_SERVER + "\",";
   json += "\"port\":" + String(MQTT_PORT) + ",";
@@ -1330,6 +1338,9 @@ void handleWiFiInfo() {
 }
 
 void handleChangeWiFi() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   server.send(200, "text/plain", "Starting WiFi Manager...");
   
   delay(1000);
@@ -1355,6 +1366,9 @@ void handleChangeWiFi() {
 }
 
 void handleSavePins() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   if (server.hasArg("sda")) SDA_PIN = server.arg("sda").toInt();
   if (server.hasArg("scl")) SCL_PIN = server.arg("scl").toInt();
   if (server.hasArg("dout")) I2S_DOUT = server.arg("dout").toInt();
@@ -1372,6 +1386,9 @@ void handleSavePins() {
 }
 
 void handleSaveMQTT() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   if (server.hasArg("server")) MQTT_SERVER = server.arg("server");
   if (server.hasArg("port")) MQTT_PORT = server.arg("port").toInt();
   if (server.hasArg("user")) MQTT_USER = server.arg("user");
@@ -1406,6 +1423,9 @@ void handleSetMode() {
 }
 
 void handleReset() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   preferences.clear();
   
   SDA_PIN = 8;
@@ -1453,6 +1473,9 @@ void handleReset() {
 }
 
 void handleRestart() {
+  if (!server.authenticate(SERVICE_USER, SERVICE_PASS)) {
+    return server.requestAuthentication();
+  }
   server.send(200, "text/plain", "Restarting...");
   delay(1000);
   ESP.restart();
